JOSHUA_MALDONADO/Flask_Angular4(backend)/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
eventlet==0.22.1
Flask==0.10.1
Flask-Bootstrap==3.1.1.2
Flask-Script==0.6.7
Flask-SocketIO==2.9.6
Flask-SQLAlchemy==1.0
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.7.2
kombu==4.1.0
MarkupSafe==0.21
python-engineio==2.0.4
python-socketio==1.9.0
pytz==2018.3
six==1.11.0
SQLAlchemy==0.9.4
vine==1.1.4
websocket-client==0.47.0
Werkzeug==0.9.4
